requirements.txt
aiogram
asyncio
dataclasses
requests
pillow